add urls module to manage all urls, and separate root layout
