+++ Frontend/einventorymanager/app/(auth)/signup/page.tsx
@@ -1,0 +1,9 @@
+import SignupForm from '@/components/auth/signup-form'
+
+export default function SignUp(){
+	return	<div className='flex justify-center align-center'>
+				<div>
+					<SignupForm />
+				</div>
+			</div>
+}
