+++ Frontend/einventorymanager/app/(root)/page.tsx
@@ -17,6 +17,7 @@ import { FaRegStar as StarIcon } from "react-icons/fa6";
 import heroImage from '@/assets/images/heroImage.png'
 import Image from "next/image";
 import AuthPopup from "@/components/others/auth-popup";
+import { login, signup, home } from '@/components/urls';
 
 
 export default function Home() {
@@ -37,7 +38,7 @@ export default function Home() {
               You can do it walking. You can do it travelling.
               You can do it on go
             </p>
-            <Link href='/sign-up' className="text-white block w-max font-medium px-6 py-2 rounded-full bg-green-600">Get started</Link>
+            <Link href={signup} className="text-white block w-max font-medium px-6 py-2 rounded-full bg-green-600">Get started</Link>
           </div>
           <div className="md:w-3/5 md:pr-0 pr-4">
               <Image src={heroImage} alt='hero image' className='md:w-3/5'/>
@@ -71,10 +72,10 @@ export default function Home() {
               </div>
               <div className="flex flex-col md:flex-row gap-4 pt-4">
                 <div className="bg-green-600 hover:bg-green-400 hover:cursor-pointer rounded-full w-full text-center px-8 py-1.5 text-white">
-                  <Link href='/login'>Login to your account</Link>
+                  <Link href={login}>Login to your account</Link>
                 </div>
                 <div className="border-2 border-green-600 hover:cursor-pointer text-green-600 hover:border-green-400 w-full text-center rounded-full px-8 py-1.5">
-                  <AuthPopup type="signup" title='Sign up for free'/>
+                  <Link href={signup}>Sign up for free</Link>
                 </div>
               </div>
             </div>
