+++ Frontend/einventorymanager/components/urls.tsx
@@ -1,0 +1,6 @@
+const home = '/'
+const login = '/login';
+const signup = '/signup';
+const contactUs = '/contact-us'
+
+export { login, signup, home, contactUs}
